Spin turns around more consistently
When a player begins a spin, it can hit something on the first few
frames and turn around, even though the spin hitboxes should destroy
it. We made hitting something in a spin try destroying it so even
when the collision box hits something, it won't turn around.

Also, grounded spin is longer than aerial spin now.