Fix problem with pre-empting camera captures running camera release after entering other capture
When two camera captures are placed next to each other, the second can
be entered, replacing the first; however, when that happens the first
capture will still run the "you exited the camera capture" function,
causing bizarre camera transition problems.

This commit adds a check to the camera capture release, to only run the
release function if the camera's capture point is the same as the
capturer says it is. If it isn't, we can assume the duties of releasing the
camera have been taken by a different camera capture node.