Disable un-invincibilty timer for hurt-flashes
When bashing was being created, an "uninvincibility" timer was added
to auto-uninvincibile the player after 0.75 seconds; however, this
got caught up in the regular hurt-recovery system, which is supposed
to be 1.6 seconds long. This caused invincibilty to be shorter than
required.

This commit parametizes the use of the uninvincibility timer, with a
default value of false, so it only starts where bashing is concerned