Disable Cutscene-lilens jumping when carrying timepieces to door
Lilens now jump over cliffs, but this causes problems
when the in the Timmy-steal-timepieces cutscene because
the lilens must fall downwards towards the door, and
jumping over it causes them to stop in midair and wiggle
side to side before falling to their goal-door. To fix this,
the jump-off-cliff condition now takes into account the
"jumping" flag in the lilen generator, so it won't do that
at points where the lilens are explicitly told not to jump.