Condense Palette Slider to double as play button
In palette selector, you press side to side to change palette, and then
press up and down to move to PLAY button and press z to play. Since
these all use different buttons, they can just be condensed to one
node that uses left, right, and z. This is done by adding a signal to
the script for the palette selector. No actual change in functionality.